rename snapshots to fixtures
