+++ tests/js/web/Notifications/fixtures/favicon.js
