+++ tests/js/web/Notifications/fixtures/faviconWithCricleNotification?radius=10&color=red.js
